+++ app/src/main/java/elite/intel/ai/brain/actions/handlers/commands/OpenGalaxyMapHandler.java
@@ -49,12 +49,7 @@ public class OpenGalaxyMapHandler implements CommandHandler {
         if (status.isInSrv()) {
             GameControllerBus.publish(new GameInputEvent(BINDING_GALAXY_MAP_BUGGY.getGameBinding(), 0));
         } if(status.isOnFoot()){
-            GameControllerBus.publish(new GameInputEvent(BINDING_ON_FOOT_WHEEL.getGameBinding(), 500));
-            GameControllerBus.publish(new GameInputEvent(BINDING_UI_LEFT.getGameBinding(), 0));
-            GameControllerBus.publish(new GameInputEvent(BINDING_UI_DOWN.getGameBinding(), 0));
-            GameControllerBus.publish(new GameInputEvent(BINDING_ACTIVATE.getGameBinding(), 0));
-            GameControllerBus.publish(new GameInputEvent(BINDING_UI_RIGHT.getGameBinding(), 0));
-            GameControllerBus.publish(new GameInputEvent(BINDING_ACTIVATE.getGameBinding(), 0));
+            GameControllerBus.publish(new GameInputEvent(BINDING_GALAXY_MAP_HUMANOID.getGameBinding(), 0));
         }
     }
 }